refactor Dashboard to be modular with components inside keeping dash simple... as well as add a shared pageHeader to be used in any tab with a header
