+++ frontend/src/components/shared/PageHeader.tsx
@@ -1,0 +1,17 @@
+interface PageHeaderProps {
+  title: string;
+  description?: string;
+  children?: React.ReactNode; // optional (for buttons or actions)
+}
+
+export function PageHeader({ title, description, children }: PageHeaderProps) {
+  return (
+    <div className="flex items-center justify-between mb-6">
+      <div>
+        <h1 className="text-2xl font-bold">{title}</h1>
+        {description && <p className="text-muted-foreground">{description}</p>}
+      </div>
+      {children && <div>{children}</div>}
+    </div>
+  );
+}
